tcga-p2/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
pandas==2.2.2
boto3==1.34.162
pymongo==4.7.3
python-dotenv==1.0.1
tqdm==4.66.4
typer==0.12.3
pandas==2.2.2
pymongo==4.7.3